- BUG: Fixed a bug related to data export. The app would crash when trying to save to the device.

+++ app/src/main/java/uk/ac/hutton/android/germinatescan/database/writer/RowDatabaseWriter.java
@@ -39,7 +39,7 @@ public class RowDatabaseWriter extends DatabaseWriter
 	@Override
 	public File write()
 	{
-		long datasetId = new PreferenceUtils(context).getInt(PreferenceUtils.PREFS_SELECTED_DATASET_ID, -1);
+		long datasetId = new PreferenceUtils(context).getLong(PreferenceUtils.PREFS_SELECTED_DATASET_ID, -1);
 		BarcodeManager barcodeManager = new BarcodeManager(context, datasetId);
 
 		List<Barcode> items = barcodeManager.getAll();
